coffee_project/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
cryptography==39.0.1
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
djwto==0.0.3
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.6.0
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.9
whitenoise==5.3.0
